[!!!][BUGFIX] Switch EXT:form hook to respect applied variants
The previously used hook `afterFormStateInitialized` is executed before
variants are applied in `FormRuntime`. This could lead to an
inconsistent state of LinkedCheckbox elements in case variants are
configured.

To circumvent this behavior, the hook `afterInitializeCurrentPage` is
now used instead.

This change is considered breaking as references to the previously
configured hook will no longer work.